display user location on the map

+++ src/c3nav/routing/locator.py
@@ -121,14 +121,14 @@ class LocatorSpace:
         self.stations = tuple(self.stations_set)
         self.stations_lookup = {station_id: i for i, station_id in enumerate(self.stations)}
 
-        self.levels = np.full((len(self.points), len(self.stations)), fill_value=-100, dtype=np.int8)
+        self.levels = np.full((len(self.points), len(self.stations)), fill_value=-100, dtype=np.int16)
         for i, point in enumerate(self.points):
             for station_id, value in point.values.items():
                 self.levels[i][self.stations_lookup[station_id]] = value
 
     def get_best_point(self, scan_values, station_ids):
         stations = tuple(self.stations_lookup[station_id] for station_id in station_ids)
-        values = np.array(tuple(scan_values[station_id] for station_id in station_ids), dtype=np.int8)
+        values = np.array(tuple(scan_values[station_id] for station_id in station_ids), dtype=np.int16)
         scores = np.sum((self.levels[:, stations]-values)**2, axis=1)
         best_point = np.argmin(scores).ravel()[0]
         return self.points[best_point], scores[best_point]
